// src/types.ts
export type Address =`0x${string}`;
export type Hex = `0x${string}`;
export type Hash = `0x${string}`;
export type EmptyHex = `0x`;
export type PromiseOrValue<T> = T | Promise<T>;
export type BigNumberish = string | bigint | number;
export type BytesLike = Uint8Array | string;


export interface SendUserOpResponse {
	userOpHash: Hash,
	transactionHash?: Hash 
}

export interface ClientOptions {
	chainId: number;
	privateKey?: string;
	rpcUrl?: string;
	isSponsoredTrx?: boolean
	noAccountCreateSponsorship?: boolean
	paymasterEndpoint?: string;
	paymaster?: "PIMLICO" | "STACKUP" | "MAGMAR";
	paymasterId?: string; //for magmar paymaster
	erc20token?: string;
	bundlerEndpoint?:string;
}

export interface TelegramClientOptions {
	chainId: number;
	telegramUserId: string;
	walletAddress: Address;
	rpcUrl?: string;
	isSponsoredTrx?: boolean
	paymasterEndpoint?: string;
	paymaster?: "PIMLICO" | "STACKUP",
	bundlerEndpoint?:string;
}